# Improve bigint `>>` performance