Avoid assigning bigint arrays from a single bigint value, which results in lot of communication since that value lives on a single locale.

16-node-cs-hdr results for a small problem size:

Before:
```
./benchmarks/run_benchmarks.py bigint_bitwise_binops -nl 16 --size=1000000
Average bigint SHIFT time = 3.8234 sec
Average bigint SHIFT rate = 0.06 GiB/sec
```

After:
```
./benchmarks/run_benchmarks.py bigint_bitwise_binops -nl 16 --size=1000000
Average bigint SHIFT time = 0.0950 sec
Average bigint SHIFT rate = 2.51 GiB/sec
```

And with a larger (default) problem size after:
```
./benchmarks/run_benchmarks.py bigint_bitwise_binops -nl 16 --size=$((10**8))
Average bigint SHIFT time = 7.4676 sec
Average bigint SHIFT rate = 3.19 GiB/sec
```